+++ R/plot.R
@@ -595,8 +595,8 @@ html_screenshot = function(x, options = opts_current$get(), ...) {
     switch(options$dev[1], pdf = '.pdf', jpeg = '.jpeg', '.png')
   } else '.png'
   wargs = options$screenshot.opts %n% list()
-  if (is.null(wargs$vwidth)) wargs$vwidth = options$out.width.px
-  if (is.null(wargs$vheight)) wargs$vheight = options$out.height.px
+  if (is.null(wargs$vwidth) && !grepl("%$", W <- options$out.width.px)) wargs$vwidth = W
+  if (is.null(wargs$vheight) && !grepl("%$", H <- options$out.height.px)) wargs$vheight = H
   if (is.null(wargs$delay)) wargs$delay = if (i1) 0.2 else 1
   d = tempfile()
   dir.create(d); on.exit(unlink(d, recursive = TRUE), add = TRUE)
@@ -609,6 +609,8 @@ html_screenshot = function(x, options = opts_current$get(), ...) {
     if (is_quarto()) "fig-format" else "dev", "' to 'png'). ",
     "See https://github.com/yihui/knitr/issues/2276 for more information."
   )
+  # set quiet opions for webshot2::webshot()
+  local_options(list(webshot.quiet = getOption('webshot.quiet', TRUE)))
   f = in_dir(d, {
     if (i1 || i3) {
       if (i1) {